requirements.txt
fastapi
requests
gspread
python-dotenv
pytz
uvicorn
gunicorn